pyproject.toml
[project]
name = "combine-backend"
version = "0.1.0"
description = "API backend for football combine testing platform"
authors = [{name = "Henrique Padueli"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "sqlmodel>=0.0.16",
    "alembic>=1.13",
    "python-multipart>=0.0.9",
    "pydantic[email]>=2.6",
    "pydantic-settings>=2.3",
    "passlib[bcrypt]>=1.7",
    "bcrypt<4",
    "python-jose[cryptography]>=3.3.0",
    "boto3>=1.34",
    "jinja2>=3.1",
    "httpx>=0.27",
    "psycopg2-binary>=2.9",
    "prometheus-fastapi-instrumentator>=6.1",
    "python-json-logger>=2.0",
    "sentry-sdk[fastapi]>=1.40",
    "opentelemetry-sdk>=1.24",
    "opentelemetry-exporter-otlp>=1.24",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.45b0",
    "opentelemetry-instrumentation-httpx>=0.45b0",
    "itsdangerous>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "httpx>=0.26",
    "ruff>=0.1.13",
    "mypy>=1.8"
]

[tool.setuptools.packages.find]
where = ["backend"]

[tool.setuptools]
package-dir = {"" = "backend"}

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
exclude = [
    "backend/statcat-storage/google-cloud-sdk/",
]